requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==23.5
requests==2.31.0
python-decouple==3.8
python-dateutil==2.8.2
pandas==2.1.4
